requirements.txt
anthropic==0.34.2
Flask==3.0.2
folium==0.16.0
gunicorn==21.2.0
haversine==2.8.1
pandas==2.2.1
python-dotenv==1.0.1
selenium==4.24.0